// tsvitereact-twcss-framermotion-web/src/shared/types.d.ts
export enum SelectedPage {
  Home = "home",
  Benefits = "benefits",
  OurClasses = "ourclasses",
  ContactUs = "contactus",
}

export interface BenefitType {
  icon: JSX.Element;
  title: string;
  description: string;
}

export interface OurClassType {
  name: string;
  description?: string;
  image: string;
}